remove _GHTTPConnection.close()
This method should *not* try to return the underlying socket back to
the socket pool as

a) it's only invoked under exceptional circumstances or when
`Connection: close' is present.  The socket pool would cull this
socket, but we can save a select(2) call if we know this socket should
be closed;

b) HTTPResponse expects to be able to share a socket with
HTTPConnection and sometimes assumes _socketobject's
reference-counting .close().  If we explicitly release the socket back
on every _GHTTPConnection.close(), we break that assumption⇧